# docs/swagger.yaml
definitions:
  dtos.AuthCodeData:
    properties:
      code:
        type: string
      email:
        type: string
    type: object
  dtos.AuthData:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  dtos.AuthResetData:
    properties:
      code:
        type: string
      email:
        type: string
      password:
        type: string
    type: object
...
  dtos.MessagesResponses:
    properties:
      message:
        type: string
    type: object
...
  dtos.UpdateGroup:
    properties:
      groups:
        type: string
    type: object
  dtos.Users:
    properties:
      email:
        type: string
    type: object
info:
  contact:
    email: cs@payuoge.com
    name: API Support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
  /auth/add-user-groups:
    post:
      consumes:
      - application/json
      description: do add user to group
      parameters:
      - description: user add to group
        in: body
        name: addUserToGroup
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateGroup'
      produces:
      - application/json
      responses:
        "200":
          description: user success to add group
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: Error Bad request
          schema:
            type: string
      security:
      - Bearer: []
      summary: Add user to group process
      tags:
      - auth
  /auth/confirm:
    post:
      consumes:
      - application/json
      description: do confirmation account
      parameters:
      - description: used for confirmation an account
        in: body
        name: confirmData
        required: true
        schema:
          $ref: '#/definitions/dtos.AuthCodeData'
      produces:
      - application/json
      responses:
        "200":
          description: redirect to home
          schema:
            type: string
        "400":
          description: Error Bad request
          schema:
            type: string
      summary: Confirmation SignUp access process
      tags:
      - auth
  /auth/forgot:
    post:
      consumes:
      - application/json
      description: do forgot password to an account
      parameters:
      - description: used for forgot password
        in: body
        name: forgotData
        required: true
        schema:
          $ref: '#/definitions/dtos.Users'
      produces:
      - application/json
      responses:
        "200":
          description: send code confirmation for forgot password
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: Error Bad request
          schema:
            type: string
      summary: ForgotPassword process
      tags:
      - auth
  /auth/google:
    get:
      consumes:
      - application/json
      description: do logout account
      produces:
      - application/json
      responses:
        "200":
          description: logged in via google successfully
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: cookie not found
          schema:
            type: string
      summary: Login google access process
      tags:
      - auth
  /auth/list-groups:
    get:
      consumes:
      - application/json
      description: do list groups
      produces:
      - application/json
      responses:
        "200":
          description: list group
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: Error Bad request
          schema:
            type: string
      security:
      - Bearer: []
      summary: list group process
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: do login account
      parameters:
      - description: login data
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/dtos.AuthData'
      produces:
      - application/json
      responses:
        "200":
          description: the message success responses
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: Error Bad request
          schema:
            type: string
      summary: Login access process
      tags:
      - auth
  /auth/logout:
    get:
      consumes:
      - application/json
      description: do logout account
      produces:
      - application/json
      responses:
        "200":
          description: logged out successfully
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: cookie not found
          schema:
            type: string
      security:
      - Bearer: []
      summary: Logout access process
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: do register account
      parameters:
      - description: used for register
        in: body
        name: registerData
        required: true
        schema:
          $ref: '#/definitions/dtos.AuthData'
      produces:
      - application/json
      responses:
        "200":
          description: send code confirmation
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: Error Bad request
          schema:
            type: string
      summary: Register access process
      tags:
      - auth
  /auth/resend:
    post:
      consumes:
      - application/json
      description: do resend code to an account
      parameters:
      - description: used for resend code
        in: body
        name: resendCode
        required: true
        schema:
          $ref: '#/definitions/dtos.Users'
      produces:
      - application/json
      responses:
        "200":
          description: check inbox email address
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: Error Bad request
          schema:
            type: string
      summary: ResendCode process
      tags:
      - auth
  /auth/reset:
    post:
      consumes:
      - application/json
      description: do reset password to an account
      parameters:
      - description: used for reset password
        in: body
        name: resetPassword
        required: true
        schema:
          $ref: '#/definitions/dtos.AuthResetData'
      produces:
      - application/json
      responses:
        "200":
          description: send for reset password
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: Error Bad request
          schema:
            type: string
      summary: ResetPassword process
      tags:
      - auth
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"